requirements.txt
protobuf==3.20.0
numpy
torch-cluster
torch-geometric
torch-scatter
torch-sparse
torch-spline-conv
tensorboard
tqdm
networkx
cnfgen
pycsp3
ijson
matplotlib
duckdb
tensorboard